poor man's push of balance and ticker updates
The server is designed only to poll plugins and does not intentionally provide any way for plugins to push updates to it. Since Coinfloor's WebSocket API is fully event-driven, polling is unnecessary and discouraged. This commit abuses the callbacks provided to the "ticker" and "balance" methods to implement a poor man's push mechanism. This has the effect that the server's internal state is updated immediately upon every balance or ticker change.